+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -7,6 +7,7 @@ import java.io.FileReader;
 import java.net.URL;
 import java.util.stream.Collectors;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.request.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
