+++ tomcat/src/main/java/org/apache/coyote/http11/HttpResponse.java
@@ -1,11 +1,20 @@
 package org.apache.coyote.http11;
 
+import java.util.LinkedHashMap;
+import java.util.Map;
+
 public class HttpResponse {
 
+    private static final String START_LINE = "HTTP/1.1 200 OK ";
+    private static final String HEADER_FORMAT_TEMPLATE = "%s: %s \r\n";
+
+    private final Map<String, String> headers = new LinkedHashMap<>();
     private final String responseBody;
 
     public HttpResponse(String responseBody) {
         this.responseBody = responseBody;
+        headers.put("Content-Type", "text/html;charset=utf-8");
+        headers.put("Content-Length", String.valueOf(responseBody.getBytes().length));
     }
 
     public byte[] getBytes() {
@@ -14,11 +23,9 @@ public class HttpResponse {
     }
 
     private String createResponse() {
-        return String.join("\r\n",
-                "HTTP/1.1 200 OK ",
-                "Content-Type: text/html;charset=utf-8 ",
-                "Content-Length: " + responseBody.getBytes().length + " ",
-                "",
-                responseBody);
+        StringBuilder sb = new StringBuilder(START_LINE).append("\r\n");
+        headers.forEach((name, value) -> sb.append(HEADER_FORMAT_TEMPLATE.formatted(name, value)));
+        sb.append("\r\n").append(responseBody);
+        return sb.toString();
     }
 }
